fix(): add conditional logic to prevent bootstrapping in protractor tests
Previously, running protractor caused a bootstrapping error because
protractor also uses NG_DEFER_BOOTSTRAP. This change monkey patches
angular.resumeBootstrap when protractor is deferring the bootstrap
so that the angular hint modules can be loaded for testing. This
fixes an already bootstrapped error received when testing with
protractor. This commit does not cause current protractor tests
to pass as the angular hint logging system is a work in progress.
The protractor tests will test for that output when the logging
is complete.

Closes #15