refactor(log): Remove log option
Now that we've switched to an event emitter, log was no longer used and was causing confusion.
The major change here is that once we emit an event, it is not stored.  If this turns out to be a big deal, we can implement a message storage module.

Closes #81

tweaked test