fix(scopes) maintain id data type
Batarang had an issue were the jsonTree would not update when
model:change events were received. This was caused by the fact
angular 1.3+ uses int ids while batarang was storing them as keys
in objects. It would pass hint the id as a string, hint would find
the scope and emit an event, but would also echo back the id as a
string. Problem is the jsonTree maintains a type correct version of
the Id, and strict comparisons then failed, hence no UI update
occurred. Updating so hint always returns the id in the type dictated
by the current angular version, regardless of the type of the id passed
to it.