fix(hint) support multi-page apps
If you visited a page on a website which did not have angular and
subsequently visited a page which did, angular would not load on it.
The second page had to be refreshed. Hint (or any other library)
signals to angular to defer bootstrapping by setting window.name
to a string starting with 'NG_DEFER_BOOTSTRAP!'. Angular then
removes this prefix for you. Because angular was not on the initial
page, the prefix was still on window.name when the second page loaded.
Hint takes the presence of this to mean protractor is running. This
causes hint to not initiate the bootstrap of angular, as protractor
would do this if it were present.

-Add a beforeunload event to remove the prefix if it is still
present when page unloads
-Update code where hint adds 'NG_DEFER_BOOTSTRAP!' to window.name
so it just add sit as a prefix to whats already there. This means the
contents
of window.name are preserved for the client application
-Update protractor detection code to just check for prefix and not do
exact equals on whole string